finalizei natives query consulta movimentos dia

+++ src/main/java/com/luizfelipe/obra/domain/dtos/ConsultaHomeObraDTO.java
@@ -1,6 +1,7 @@
 package com.luizfelipe.obra.domain.dtos;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -8,17 +9,17 @@ public class ConsultaHomeObraDTO implements Serializable {
 
 	private static final long serialVersionUID = 1L;
 
-	private Long idObra;
+	private BigInteger idObra;
 
 	private String obra;
 
 	private List<ConsultaHomeObraItemDTO> itens = new ArrayList<ConsultaHomeObraItemDTO>();
 
-	public Long getIdObra() {
+	public BigInteger getIdObra() {
 		return idObra;
 	}
 
-	public void setIdObra(Long idObra) {
+	public void setIdObra(BigInteger idObra) {
 		this.idObra = idObra;
 	}
